requirements.txt
# ── wildfire-uncertainty-gnn — pip requirements (pinned) ──────────────────
# Use environment.yml + conda for the full install (handles GDAL/rasterio).
# This file is for CI, Docker, or CPU-only environments.
#
# Install order matters for PyG:
#   pip install torch==2.1.2 --index-url https://download.pytorch.org/whl/cu118
#   pip install torch-geometric torch-scatter torch-sparse ... (see README)
#   pip install -r requirements.txt

# ── PyTorch Geometric ──────────────────────────────────────────────────────
torch-geometric==2.5.2

# ── Geospatial ─────────────────────────────────────────────────────────────
rasterio==1.3.9
geopandas==0.14.3
fiona==1.9.5
pyproj==3.6.1
shapely==2.0.3

# ── Scientific core ────────────────────────────────────────────────────────
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.2
pandas==2.2.1
pyarrow==15.0.2

# ── Visualization ──────────────────────────────────────────────────────────
matplotlib==3.8.3
seaborn==0.13.2

# ── Gradient boosting baselines ───────────────────────────────────────────
xgboost==2.0.3
lightgbm==4.3.0

# ── Calibration ───────────────────────────────────────────────────────────
netcal==1.3.5

# ── Config management ─────────────────────────────────────────────────────
pyyaml==6.0.1
omegaconf==2.3.0

# ── ML experiment tracking ────────────────────────────────────────────────
mlflow==2.11.3

# ── Utilities ────────────────────────────────────────────────────────────
tqdm==4.66.2
rich==13.7.1
python-dotenv==1.0.1

# ── Dev / test ────────────────────────────────────────────────────────────
black==24.3.0
ruff==0.3.5
pytest==8.1.1
pytest-cov==5.0.0